+++ src/spec/mod.rs
@@ -209,9 +209,9 @@ impl Spec {
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-
     use pretty_assertions::assert_eq;
+
+    use super::*;
 
     #[test]
     fn spec_extensions_deserialize() {